pyproject.toml
[tool.poetry]
name = "geniac-collect-jp"
version = "0.1.0"
description = ""
authors = ["Eiji Takahashi <comu2e@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
warcio = "^1.7.4"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
tqdm = "^4.66.2"
boto3 = "^1.34.54"
lxml = "^5.1.0"
google-auth = "^2.29.0"
pydantic = "^2.7.1"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
pg8000 = "^1.31.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"